Improved: Show TO filters if the facet has orders; disable if none.(#673)
